feat: add input doc generation helper
This adds an optional helper for generating an input document from a set
of keys and services. This helper is intentionally simple; it won't
cover all use cases. But when all you need is a basic document, it will
get you there without having to worry about contexts and such.

Additionally, those familiar with the did-peer-2 library will see
similarities between the helper added here and the `generate` method in
the did-peer-2 library. This is deliberate and should hopefully help
ease the transition for those adopting did:peer:4 after did:peer:2.

I accidentally made the generation of did:peer:2 actually pretty okay,
interface wise. This addition seeks to prevent that from becoming a
barrier to adoption.

Signed-off-by: Daniel Bluhm <dbluhm@pm.me>